// src/components/campaigns/campaigns.interface.ts
export interface CampaignsTableProps {
  profileImage: string;
  name: string;
  id: string;
  email: string;
  timeAdded: string;
  countryFlag: string;
  goferId: string
}

export interface CampaignsProductTableProps {
  productImage: string;
  productName: string;
  categories: string;
  status: string;
}